+++ src/app/service/index.ts
@@ -1,3 +1,5 @@
 export { IpaService } from './ipa.service';
 export { InventoryService } from './inventory.service';
+export { ParserService } from './parser.service';
+export { PermutatorService } from './permutator.service';
 export { TranscriptionService } from './transcription.service';
